refactor(go-client): move selectPodIP to ip.go and normalize parsed IP
